Move GetCRCLatestVersionFromMirror to 'bundle' package
Having GetCRCLatestVersionFromMirror in the 'version' package is
problametic, as this code is fairly high level (it downloads from the
net and needs the 'download' package), while the rest of 'version' is
dedicated to provided build time constants for use by other parts of
the code.

In particular, having GetCRCLatestVersionFromMirror in 'version'
prevents the 'download' package from using 'version', which I need to do
in the following commits.

It's moved to pkg/crc/machine/bundle for now, but it's not a great fit
there.

I've also renamed the function to FetchLatestReleaseInfo() to make the
name a bit shorter.
